requirements.txt
python-docx>=1.1.0
Pillow>=10.0.0
sympy>=1.12
